requirements.txt
# Google APIs for Gmail and Sheets
google-api-python-client==2.111.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
gspread==5.12.0
oauth2client==4.1.3

# For parsing and configuration
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pandas==2.1.4

# For development and debugging
structlog==23.2.0 